+++ package.json
@@ -16,7 +16,7 @@
     "@hookform/resolvers": "^3.9.0",
     "@oslojs/jwt": "^0.2.0",
     "@oslojs/oauth2": "^0.5.0",
-    "@prisma/client": "6.1.0",
+    "@prisma/client": "6.2.1",
     "@radix-ui/react-alert-dialog": "^1.1.4",
     "@radix-ui/react-avatar": "^1.0.4",
     "@radix-ui/react-checkbox": "^1.1.3",
@@ -78,7 +78,7 @@
     "prettier": "^3.2.5",
     "prettier-plugin-sort-imports": "^1.8.6",
     "prettier-plugin-tailwindcss": "^0.6.6",
-    "prisma": "^6.1.0",
+    "prisma": "^6.2.1",
     "tailwindcss": "^3.3.0",
     "typescript": "^5"
   },
